Add a FlowInvalidInputTest suppressor comment
This can be used to allow tests that pass invalid data to a function
etc. Even if styled-components use flow it is not given that the users
will so it is still necessary to have tests that make sure it does not
break on runtime with bad data even though flow errors on this on
compile time.