Add first support for the private API
To add support for server-side rendering, it needs to be possible to
pass in an identifier from the outside. This is a private API because
the only thing ever writing this is babel-plugin-styled-components,
which will automatically turn your normal styled components into the ssr
compatible version by turning your styled() calls into this notation:

    styled({ target: 'div', identifier: 'YourComponent-asdf123', displayName: 'YourComponent' })

This commit adds support for this notation.